+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED] - 2022-##-##
+### Added
+- plex show search button
+- plex search user pref
+
+### Changed
+- svgr config retains viewboxes
 
 ## [1.18.0] - 2022-05-07
 ### Added